FEAT: delete user , add user

+++ frontend/src/components/DeleteUserModel.jsx
@@ -2,46 +2,63 @@
 /* eslint-disable react/prop-types */
 
 import axios from "axios";
-import React from "react";
+import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import { deleteUserApi } from "../constants/apiUrls";
 import { ImCross } from "react-icons/im";
 import { toast } from "react-toastify";
+import Spinner from "./Spinner";
 
 const DeleteUserModel = ({ userId, setIsDeleteUserModelOpen, setNeedReload }) => {
+	const [isLoading, setIsLoading] = useState(false);
+
 	const loggedInUser = useSelector((store) => store.user.userDetails);
 	// console.log(loggedInUser);
+
 	// console.log(userId);
 
 	const handledDeleteUser = async () => {
-		console.log("Deleting......");
+		try {
+			if (!loggedInUser.role === "admin") {
+				return toast.warn(`You are not an admin !`);
+			}
+			setIsLoading(true);
 
-		let headersList = {
-			Accept: "*/*",
-			Authorization: `Bearer ${loggedInUser.token}`,
-			"Content-Type": "application/json",
-		};
+			// console.log("Deleting......");
+			let headersList = {
+				Accept: "*/*",
+				Authorization: `Bearer ${loggedInUser.token}`,
+				"Content-Type": "application/json",
+			};
 
-		let bodyContent = JSON.stringify({
-			userId,
-		});
+			let bodyContent = JSON.stringify({
+				userId,
+			});
 
-		let reqOptions = {
-			url: deleteUserApi,
-			method: "DELETE",
-			headers: headersList,
-			data: bodyContent,
-		};
+			let reqOptions = {
+				url: deleteUserApi,
+				method: "DELETE",
+				headers: headersList,
+				data: bodyContent,
+			};
 
-		let response = await axios.request(reqOptions);
-		// console.log(response);
+			let response = await axios.request(reqOptions);
+			// console.log(response);
 
-		if (!response.data.success) {
-			return toast.error(response.data.message || "User Deletion Failed");
+			if (!response.data.success) {
+				return toast.error(response.data.message || "User Deletion Failed");
+			}
+
+			setNeedReload((prev) => !prev);
+			setIsLoading(false);
+			return toast.success(
+				`${response.data?.message || "Successfully deleted successfully!"}`
+			);
+		} catch (error) {
+			setIsLoading(false);
+			console.log(error);
+			return toast.error(`${error.data?.message || "User Deletion Failed !"}`);
 		}
-
-		setNeedReload((prev) => !prev);
-		return toast.success(`${response.data?.message || "Successfully deleted successfully!"}`);
 	};
 
 	return (
@@ -89,10 +106,16 @@ const DeleteUserModel = ({ userId, setIsDeleteUserModelOpen, setNeedReload }) =>
 								Are you sure you want to delete this user?
 							</h3>
 							<button
-								className='text-white bg-red-600 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-base inline-flex items-center px-3 py-2.5 text-center mr-2 dark:focus:ring-red-800'
+								className='text-white bg-red-600 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-base inline-flex items-center px-3 py-2.5 text-center mr-2 dark:focus:ring-red-800 justify-center disabled:cursor-not-allowed'
 								onClick={handledDeleteUser}
+								disabled={isLoading}
 							>
-								Yes, I'am sure
+								<span className='ml-1'>Yes, I'am sure</span>
+								{isLoading && (
+									<span className='mt-0.5 ml-1'>
+										<Spinner />
+									</span>
+								)}
 							</button>
 							<button
 								type='button'
